move model deployment into own file to have better dependency separation
